+++ rust/lance/src/dataset/write.rs
@@ -142,8 +142,7 @@ pub async fn write_fragments(
 
         if num_rows_in_current_file >= params.max_rows_per_file {
             let num_rows = writer.take().unwrap().finish().await?;
-            // TODO: do we need these both?
-            debug_assert_eq!(num_rows, num_rows_in_current_file as u64);
+            debug_assert_eq!(num_rows, num_rows_in_current_file);
             fragments.last_mut().unwrap().fragment_length = num_rows;
             num_rows_in_current_file = 0;
         }
